+++ src/core/providers/helpers/ledger/components/LedgerModalComponent/helpers/getAuthTokenText.ts
