+++ src/core/providers/helpers/ledger/components/LedgerModalComponent/ldegerModalComponent.styles.ts
